// frontend/src/api/generated/models/ArtifactData.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Source location for ingesting an artifact.
 *
 * Provide a single downloadable url pointing to a bundle that contains the artifact assets.
 */
export type ArtifactData = {
    /**
     * Artifact source url used during ingest.
     */
    url: string;
};